move interpolate_from_lower_zoom
